test(generator): fold benchmark into playground integration
