+++ packages/generator/__tests__/integration-playground.test.ts
@@ -2,8 +2,7 @@ import generate from '../src/generate.js';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { Template } from '@pdfme/common';
-import { getInputFromTemplate } from '@pdfme/common';
+import { BLANK_PDF, getInputFromTemplate, Template, type Schema } from '@pdfme/common';
 import {
   text,
   image,
@@ -22,6 +21,51 @@ import { getFont, getImageSnapshotOptions, pdfToImages } from './utils.js';
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const PERFORMANCE_THRESHOLD = parseFloat(process.env.PERFORMANCE_THRESHOLD || '1.5');
+const benchmarkTest = process.env.PDFME_GENERATOR_BENCHMARK ? test : test.skip;
+const generatorPlugins = {
+  text,
+  image,
+  svg,
+  line,
+  rectangle,
+  ellipse,
+  signature,
+  table,
+  list,
+  multiVariableText,
+  ...barcodes,
+};
+
+const measureGenerate = async (name: string, fn: () => Promise<void>) => {
+  const runs: number[] = [];
+  for (let i = 0; i < 6; i += 1) {
+    const startedAt = process.hrtime.bigint();
+    await fn();
+    runs.push(Number(process.hrtime.bigint() - startedAt) / 1000000);
+  }
+
+  const measuredRuns = runs.slice(1);
+  const average =
+    measuredRuns.reduce((total, duration) => total + duration, 0) / measuredRuns.length;
+  const sortedRuns = [...measuredRuns].sort((a, b) => a - b);
+  const median = sortedRuns[Math.floor(sortedRuns.length / 2)];
+  const variance =
+    measuredRuns.reduce((total, duration) => total + (duration - average) ** 2, 0) /
+    measuredRuns.length;
+  const standardDeviation = Math.sqrt(variance);
+
+  console.info(
+    `[@pdfme/generator benchmark] ${name}: ${average.toFixed(2)}ms average, ` +
+      `${median.toFixed(2)}ms median, ${standardDeviation.toFixed(2)}ms stddev after warmup`,
+  );
+  console.info(
+    `[@pdfme/generator benchmark] ${name} runs: ${runs
+      .map((duration) => duration.toFixed(2))
+      .join(', ')}ms`,
+  );
+
+  expect(average).toBeGreaterThan(0);
+};
 
 // Load all templates from playground/public/template-assets
 function loadPlaygroundTemplates(): Record<string, Template> {
@@ -90,19 +134,7 @@ describe('generate integration test(playground)', () => {
         const pdf = await generate({
           inputs,
           template,
-          plugins: {
-            text,
-            image,
-            svg,
-            line,
-            rectangle,
-            ellipse,
-            signature,
-            table,
-            list,
-            multiVariableText,
-            ...barcodes,
-          },
+          plugins: generatorPlugins,
           options: { font },
         });
 
@@ -123,4 +155,53 @@ describe('generate integration test(playground)', () => {
       });
     }
   });
+
+  benchmarkTest(
+    'benchmarks generator-only paths used by performance-sensitive templates',
+    async () => {
+      const font = getFont();
+      const textSchemas: Schema[] = Array.from({ length: 120 }, (_, index) => ({
+        name: `field${index}`,
+        type: 'text',
+        content: '',
+        position: { x: 10 + (index % 4) * 45, y: 10 + Math.floor(index / 4) * 8 },
+        width: 40,
+        height: 6,
+        fontSize: 8,
+      }));
+      const textHeavyTemplate: Template = { basePdf: BLANK_PDF, schemas: [textSchemas] };
+      const textHeavyInputs = Array.from({ length: 25 }, (_, inputIndex) =>
+        Object.fromEntries(
+          textSchemas.map((schema, schemaIndex) => [
+            schema.name,
+            `${inputIndex}-${schemaIndex}`,
+          ]),
+        ),
+      );
+
+      await measureGenerate('blank text-heavy template', () =>
+        generate({ inputs: textHeavyInputs, template: textHeavyTemplate, options: { font } }),
+      );
+
+      const labelTemplate = playgroundTemplates['address-label-30'];
+      if (!labelTemplate) {
+        throw new Error('Failed to load playground template "address-label-30".');
+      }
+      const [baseLabelInput = {}] = getInputFromTemplate(labelTemplate);
+      const labelInputs = Array.from({ length: 40 }, (_, index) => ({
+        ...baseLabelInput,
+        '{1}Name': `Kyohei Fukuda ${index}`,
+      }));
+
+      await measureGenerate('playground address-label-30 multi-input template', () =>
+        generate({
+          inputs: labelInputs,
+          template: labelTemplate,
+          plugins: generatorPlugins,
+          options: { font },
+        }),
+      );
+    },
+    120000,
+  );
 });
